Refactor MTE-2557 SettingsAppearanceTest and TestRail (#19946)
* empty commit

* Add TestRail URL. Uncomment currently working tests.

* Missing one URL

* remove print statement

* Align tests from WebsiteAccessTests to TestRail

* Move WebsiteAccessTest to SettingAppearanceTest

* Remove WebsiteAccessTest.swift

* Rename SettingAppearanceTest to SettingTest

* Add TestRail test for testVisitWebiste
