+++ focus-ios/focus-ios-tests/SmokeTest.xctestplan
@@ -20,6 +20,7 @@
         "DragAndDropTest",
         "FindInPageTest\/testFindInPageURLBarElement()",
         "OpenInFocusTest\/testOpenViaSafari()",
+        "PageShortcutsTest\/testAddRemoveShortcut()",
         "PastenGoTest\/testClipboard()",
         "QuickAddAutocompleteURLTest",
         "SearchProviderTest\/testAddRemoveCustomSearchProvider()",
@@ -27,10 +28,12 @@
         "SearchProviderTest\/testPreventionOfRemovingDefaultSearchProvider()",
         "SearchProviderTest\/testWikipediaSearchProvider()",
         "SearchSuggestionsPromptTest",
-        "SettingAppearanceTest\/testAddRemoveCustomDomain()",
-        "SettingAppearanceTest\/testDisableAutocomplete()",
-        "SettingAppearanceTest\/testOpenInSafari()",
-        "PageShortcutsTest\/testAddRemoveShortcut()",
+        "SettingTest\/testAddRemoveCustomDomain()",
+        "SettingTest\/testAutocompleteCustomDomain()",
+        "SettingTest\/testDisableAutocomplete()",
+        "SettingTest\/testEnableDisableAutocomplete()",
+        "SettingTest\/testOpenInSafari()",
+        "SettingTest\/testReEnableAutoComplete()",
         "TrackingProtectionTest\/testInactiveSettings()",
         "UserAgentTest",
         "WebsiteAccessTests\/testAutocompleteCustomDomain()",
